requirements.txt
asgiref==3.2.7
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.0.1
django-environ==0.4.5
django-allauth==0.41.0
django-extensions==2.2.9
django-filter==2.2.0
django-rest-auth==0.9.5
drf-yasg==1.17.1
django-model-utils==4.0.0